pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jesterTOV"]

[project]
name = "jesterTOV"
authors = [
  {name = "Thibeau Wouters", email = "thibeauwouters@gmail.com"},
  {name = "Peter T. H. Pang", email = "tsunhopang@gmail.com"}
]
version = "0.1.1"
description = "JAX-accelerated nuclear equation of state code and TOV solver with automatic differentiation"
keywords = ["neutron-stars", "eos", "tov", "jax", "physics", "astronomy", "automatic-differentiation"]
readme = "README.md"
requires-python = ">=3.10, <3.13"
license = {file = "LICENSE"}
urls = {repository = "https://github.com/nuclear-multimessenger-astronomy/jester"}
dependencies = [
    "jax>=0.4.23",
    "jaxtyping>=0.2.24",
    "typing_extensions>=4.5.0",
    "diffrax>=0.5.0",
    "interpax",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "equinox>=0.11.0",
    "flowjax>=12.0.0",
    "flowmc>=0.3.4",
    "beartype>=0.10.0",
    "matplotlib>=3.10.7",
    "seaborn>=0.12.0",
    "corner>=2.2.0",
    "scipy>=1.10.0",
    "tqdm>=4.65.0",
    "zenodo-get>=1.5.0"
]

classifiers = [
    "Programming Language :: Python :: 3",
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'License :: OSI Approved :: MIT License',
    'Intended Audience :: Science/Research',
    'Intended Audience :: End Users/Desktop',
    'Intended Audience :: Developers',
    'Natural Language :: English',
    'Topic :: Scientific/Engineering',
    'Topic :: Scientific/Engineering :: Astronomy',
    'Topic :: Scientific/Engineering :: Physics',
    'Operating System :: POSIX',
    'Operating System :: Unix',
    'Operating System :: MacOS',
]

[project.scripts]
run_jester_inference = "jesterTOV.inference.run_inference:cli_entry_point"
run_jester_postprocessing = "jesterTOV.inference.postprocessing.postprocessing:main"

[project.optional-dependencies]
test = ["pytest>=7.0", "pytest-cov>=4.0", "pytest-xdist>=3.0", "hypothesis>=6.0"]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "pytest-xdist>=3.0", "hypothesis>=6.0", "pre-commit>=4.5.0", "bilby>=2.0.0", "astropy>=5.0.0"]
examples = ["ipykernel>=7.1.0"]
docs = [
    "sphinx>=7.0",
    "sphinx-book-theme>=1.0",
    "nbsphinx>=0.9",
    "sphinx-autodoc-typehints>=1.12",
    "sphinx-copybutton>=0.5",
    "myst-parser>=2.0"
]

[tool.ruff]
extend-exclude = ["examples/"]

[tool.ruff.lint]
ignore = [
    "F722",  # Forward annotation false positive
    "F821",  # Undefined name (jaxtyping dimension names)
    "E731",  # Lambda expressions
    "E741",  # Ambiguous variable name
    "F841",  # Unused local variable
    "E722",  # Bare except
    "E402",  # Module level import not at top of file
]

[tool.ruff.lint.isort]
combine-as-imports = true
extra-standard-library = ["typing_extensions"]
lines-after-imports = 2
order-by-type = false

[tool.pyright]
reportIncompatibleMethodOverride = true
include = ["jesterTOV"]

# pyproject.toml
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.coverage.run]
source = ["jesterTOV"]
omit = ["*/tests/*", "*/examples/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]
